# stdlib: Fix mime_decode/1 binary matching performance