Symptom: Throughput of base64:mime_decode/1 significantly lower than
base64:decode/1.

Problem: tail_contains_more/2 prevents compiler from delaying creation
of sub binaries.

Solution: Restructure mime_decode_binary/2 to use binary matching best
practices from the Efficiency Guide.

See ERL-366

---

# Problem
Using erlc +bin_opt_info on the base64 module, there is a warning about tail_contains_more/2

> lib/stdlib/src/base64.erl:218: Warning: OPTIMIZED: creation of sub binary delayed
> lib/stdlib/src/base64.erl:220: Warning: NOT OPTIMIZED: called function tail_contains_more/2 does not begin with a suitable binary matching instruction

# Measurements
Based on my testing, I verified that this does pose a problem, at least for large binaries (I used one that is 400000 bytes.

    Function               | Avg exec | Throughput
    -----------------------------------------------
    base64:decode/1        | 0.017 s  | 22.502 MiB/s
    base64:mime_decode/1   | 0.029 s  | 12.953 MiB/s

With this pull request, we can now match the throughput of decode/1:

    Function               | Avg exec | Throughput
    -----------------------------------------------
    base64:decode/1        | 0.018 s  | 21.170 MiB/s
    base64:mime_decode/1   | 0.016 s  | 23.519 MiB/s

The module to produce these measurements is included in ERL-366.

# Solution Description
My proposed fix basically merges the functionality of tail_contains_more/2 and mime_decode_binary/2 into a second-stage decode step called mime_decode_binary_after_eq/3. I believe the implementation is functionally equivalent, with the added benefit allowing the compiler to delay sub binary creation (bin_opt_info).

To see how the two are merged, viewing the diff without whitespace helps a little bit, but while preparing this PR, I had to view the entire base64 files side by side (old/new). This allowed me to jump down to tail_contains_more/2 more easily while comparing. I hope this helps the review process.

erlc +bin_opt_info lib/stdlib/src/base64.erl:
> lib/stdlib/src/base64.erl:218: Warning: OPTIMIZED: creation of sub binary delayed
> lib/stdlib/src/base64.erl:220: Warning: OPTIMIZED: creation of sub binary delayed
> lib/stdlib/src/base64.erl:233: Warning: OPTIMIZED: creation of sub binary delayed
> lib/stdlib/src/base64.erl:235: Warning: OPTIMIZED: creation of sub binary delayed

# Tests
I ran the tests for base64_SUITE -- all good.

I ran the tests for stdlib -- I had **2 failures on the otp maint branch**, excluding the changes from this pull request. I don't think they're relevant to this PR, but I wanted to provide them for completeness.

This one, from calendar_SUITE seems to be a problem with my local time settings:
```
=== Ended at 2017-02-24 11:27:27
=== Location: [{calendar_SUITE,local_time_to_universal_time_dst_x,129},
              {test_server,ts_tc,1529},
              {test_server,run_test_case_eval1,1045},
              {test_server,run_test_case_eval,977}]
=== Reason: no match of right hand side value {{2003,1,15},{19,0,0}}
  in function  calendar_SUITE:local_time_to_universal_time_dst_x/1 (calendar_SUITE.erl, line 129)
  in call from test_server:ts_tc/3 (test_server.erl, line 1529)
  in call from test_server:run_test_case_eval1/6 (test_server.erl, line 1045)
  in call from test_server:run_test_case_eval/9 (test_server.erl, line 977)
```

I don't have any idea about this one:
```
=== Ended at 2017-02-24 11:37:46
=== Location: [{shell_SUITE,'-scan/1-fun-0-',2832},
              {shell_SUITE,scan,2841},
              {shell_SUITE,records,401},
              {test_server,ts_tc,1529},
              {test_server,run_test_case_eval1,1045},
              {test_server,run_test_case_eval,977}]
=== Reason: no match of right hand side value 
                 {error,{1,erl_parse,["syntax error before: ","'*'"]}}
  in function  shell_SUITE:'-scan/1-fun-0-'/1 (shell_SUITE.erl, line 2832)
  in call from shell_SUITE:scan/2 (shell_SUITE.erl, line 2841)
  in call from shell_SUITE:records/1 (shell_SUITE.erl, line 401)
  in call from test_server:ts_tc/3 (test_server.erl, line 1529)
  in call from test_server:run_test_case_eval1/6 (test_server.erl, line 1045)
  in call from test_server:run_test_case_eval/9 (test_server.erl, line 977)
```

Thanks for reading!