@bjorng 2017-02-27
Thanks, looks good.

Just one more thing... Please squash your correction commit into the original commit. When you have done that I'll add the PR to our daily builds.